+++ Realm/ObjectServerTests/RLMSyncTestCase.mm
@@ -613,7 +613,7 @@ static bool s_opensApp;
     RLMApp *app = [RLMApp appWithConfiguration:config];
     RLMSyncManager *syncManager = app.syncManager;
     syncManager.userAgent = self.name;
-    RLMLogger.defaultLogger.level = RLMLogLevelWarn;
+    [RLMLogger setLevel:RLMLogLevelWarn forCategory:RLMLogCategorySync];
     return app;
 }